Tests for all signature schemes
